Moved wedneday logic into a separate file
